requirements.txt
asttokens==2.4.1
contourpy==1.2.1
cycler==0.12.1
decorator==5.1.1
dill==0.3.8
executing==2.0.1
fonttools==4.51.0
iniconfig==2.0.0
ipython==8.23.0
jedi==0.19.1
kiwisolver==1.4.5
matplotlib-inline==0.1.7
mpmath==1.3.0
multiprocess==0.70.16
numpy==1.26.4
packaging==24.0
parso==0.8.4
pexpect==4.9.0
pillow==10.3.0
pluggy==1.4.0
prompt-toolkit==3.0.43
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.17.2
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
six==1.16.0
stack-data==0.6.3
sympy==1.12
traitlets==5.14.2
typing_extensions==4.11.0
urllib3==2.2.1
wcwidth==0.2.13
